\name{CONTundirv6-package}
\alias{CONTundirv6-package}
\alias{CONTundirv6}
\docType{package}
\title{
\packageTitle{CONTundirv6}
}
\description{
\packageDescription{CONTundirv6}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CONTundirv6}
\packageIndices{CONTundirv6}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CONTundirv6}

Maintainer: \packageMaintainer{CONTundirv6}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
